API documentation

API documentation

Index

Exported Functions

Internal Functions

Index

TermStructureModels.Forecast

TermStructureModels.Hyperparameter

TermStructureModels.LatentSpace

TermStructureModels.Parameter

TermStructureModels.PosteriorSample

TermStructureModels.ReducedForm

TermStructureModels.Scenario

TermStructureModels.TermPremium

TermStructureModels.YieldCurve

Base.getindex

Base.getindex

Statistics.mean

Statistics.median

Statistics.quantile

Statistics.std

Statistics.var

TermStructureModels.AR_res_var

TermStructureModels.Aₓ

TermStructureModels.Aₚ

TermStructureModels.Bₓ

TermStructureModels.Bₚ

TermStructureModels.GQ_XX

TermStructureModels.Kϕ

TermStructureModels.LDL

TermStructureModels.Minnesota

TermStructureModels.NIG_NIG

TermStructureModels.PCA

TermStructureModels.PCs_2_latents

TermStructureModels.S

TermStructureModels.S_hat

TermStructureModels.T0P

TermStructureModels.T1X

TermStructureModels._conditional_forecasts

TermStructureModels._scenario_analysis

TermStructureModels._scenario_analysis_unconditional

TermStructureModels._termPremium

TermStructureModels._unconditional_forecasts

TermStructureModels.aτ

TermStructureModels.bτ

TermStructureModels.calibrate_μϕ_const

TermStructureModels.conditional_forecasts

TermStructureModels.dcurvature_dτ

TermStructureModels.dimQ

TermStructureModels.erase_nonstationary_param

TermStructureModels.fitted_YieldCurve

TermStructureModels.generative

TermStructureModels.isstationary

TermStructureModels.jensens_inequality

TermStructureModels.latentspace

TermStructureModels.log_marginal

TermStructureModels.loglik_mea

TermStructureModels.loglik_tran

TermStructureModels.post_kQ_infty

TermStructureModels.post_Σₒ

TermStructureModels.post_γ

TermStructureModels.post_κQ

TermStructureModels.post_ϕ_σ²FF

TermStructureModels.posterior_sampler

TermStructureModels.prior_C

TermStructureModels.prior_γ

TermStructureModels.prior_κQ

TermStructureModels.prior_σ²FF

TermStructureModels.prior_ϕ0

TermStructureModels.reducedform

TermStructureModels.scenario_analysis

TermStructureModels.term_premium

TermStructureModels.tuning_hyperparameter

TermStructureModels.yϕ_Xϕ

TermStructureModels.ν

TermStructureModels.ϕ_2_ϕ₀_C

TermStructureModels.ϕ_hat

TermStructureModels.ϕ_σ²FF_2_ΩFF

TermStructureModels.ϕ_σ²FF_2_ΩPP

Exported Functions

TermStructureModels.Forecast — Type

@kwdef struct Forecast <: PosteriorSample

It contains a result of the scenario analysis, the conditional prediction for yields, factors = [PCs macros], and term premiums.

yields

factors

TP: term premium forecasts

TermStructureModels.Hyperparameter — Type

@kwdef struct Hyperparameter

p::Int

q::Matrix

ν0

Ω0::Vector

μϕ_const::Vector = zeros(length(Ω0)): It is a prior mean of a constant term in our VAR.

TermStructureModels.LatentSpace — Type

@kwdef struct LatentSpace <: PosteriorSample

When the model goes to the JSZ latent factor space, the statistical parameters in struct Parameter are also transformed. This struct contains the transformed parameters. Specifically, the transformation is latents[t,:] = T0P_ + inv(T1X)*PCs[t,:].

In the latent factor space, the transition equation is data[t,:] = KₚXF + GₚXFXF*vec(data[t-1:-1:t-p,:]') + MvNormal(O,ΩXFXF), where data = [latent macros].

latents::Matrix

κQ

kQ_infty

KₚXF::Vector

GₚXFXF::Matrix

ΩXFXF::Matrix

TermStructureModels.Parameter — Type

@kwdef struct Parameter <: PosteriorSample

It contains statistical parameters of the model that are sampled from function posterior_sampler.

κQ::Float64

kQ_infty::Float64

ϕ::Matrix{Float64}

σ²FF::Vector{Float64}

Σₒ::Vector{Float64}

γ::Vector{Float64}

TermStructureModels.PosteriorSample — Type

abstract type PosteriorSample

It is a super-set of structs Parameter, ReducedForm, LatentSpace, YieldCurve, TermPremium, Forecast.

TermStructureModels.ReducedForm — Type

@kwdef struct ReducedForm <: PosteriorSample

It contains statistical parameters in terms of the reduced form VAR(p) in P-dynamics. λP and ΛPF are parameters in the market prices of risks equation, and they only contain the first dQ non-zero equations.

κQ

kQ_infty

KₚF

GₚFF

ΩFF::Matrix

Σₒ::Vector

λP

ΛPF

mpr::Matrix(market prices of risks, T, dP)

TermStructureModels.Scenario — Type

@kwdef struct Scenario

It contains scenarios to be conditioned in the scenario analysis. When y = [yields; macros] is a observed vector in our measurement equation, Scenario.combinations*y = Scenario.values constitutes the scenario at a specific time. Vector{Scenario} is used to describe a time-series of scenarios.

combinations and values should be Matrix and Vector. If values is a scalar, combinations would be a matrix with one raw vector and values should be one-dimensional vector, for example [values].

combinations::Matrix

values::Vector

TermStructureModels.TermPremium — Type

@kwdef struct TermPremium <: PosteriorSample

It contains a estimated time series of a term premium for one maturity.

TP::Vector: term premium estimates of a specific maturity bond. TP = timevarying_TP + const_TP + jensen holds.

timevarying_TP::Matrix: rows:time, cols:factors, values: contributions of factors on TP

const_TP::Float64: constant part in TP

jensen::Float64: the part due to the Jensen's inequality

TermStructureModels.YieldCurve — Type

@kwdef struct YieldCurve <: PosteriorSample

It contains a fitted yield curve. yields[t,:] = intercept + slope*latents[t,:] holds.

latents::Matrix: latent pricing factors in LatentSpace

yields

intercept

slope

Base.getindex — Method

getindex(x::PosteriorSample, c::Symbol)

For struct <: PosteriorSample, struct[:name] calls objects in struct.

Base.getindex — Method

getindex(x::Vector{<:PosteriorSample}, c::Symbol)

For struct <: PosteriorSample, struct[:name] calls objects in struct. Output[i] = $i'$th posterior sample

Statistics.mean — Method

mean(x::Vector{<:PosteriorSample})

Output[:variable name] returns the corresponding posterior mean.

Statistics.median — Method

median(x::Vector{<:PosteriorSample})

Output[:variable name] returns the corresponding posterior median.

Statistics.quantile — Method

quantile(x::Vector{<:PosteriorSample}, q)

Output[:variable name] returns a quantile of the corresponding posterior distribution.

Statistics.std — Method

std(x::Vector{<:PosteriorSample})

Output[:variable name] returns the corresponding posterior standard deviation.

Statistics.var — Method

var(x::Vector{<:PosteriorSample})

Output[:variable name] returns the corresponding posterior variance.

TermStructureModels.AR_res_var — Method

AR_res_var(TS::Vector, p)

It derives an MLE error variance estimate of an AR(p) model

Input

univariate time series TS and the lag p

output(2)

residual variance estimate, AR(p) coefficients

TermStructureModels.GQ_XX — Method

GQ_XX(; κQ)

κQ governs a conditional mean of the Q-dynamics of X, and its slope matrix has a restricted form. This function shows that restricted form.

Output

slope matrix of the Q-conditional mean of X

TermStructureModels.LDL — Method

LDL(X)

This function generate a matrix decomposition, called LDLt. X = L*D*L', where L is a lower triangular matrix and D is a diagonal. How to conduct it can be found at Wikipedia.

Input

Decomposed Object, X

Output(2)

L, D

Decomposed result is X = L*D*L'

TermStructureModels.PCA — Function

PCA(yields, p, proxies=[]; rescaling=false)

It derives the principal components from yields.

Input

yields[p+1:end, :] is used to construct the affine transformation, and then all yields[:,:] are transformed into the principal components.

Since signs of PCs is not identified, we use proxies to identify the signs. We flip PCs to make cor(proxies[:, i]. PCs[:,i]) > 0. If proxies is not given, we use the following proxies as a default: [yields[:, end] yields[:, end] - yields[:, 1] 2yields[:, Int(floor(size(yields, 2) / 3))] - yields[:, 1] - yields[:, end]].

size(proxies) = (size(yields[p+1:end, :], 1), dQ)

If rescaling == true, all PCs and OCs are normalized to have an average std of yields.

Output(4)

PCs, OCs, Wₚ, Wₒ, mean_PCs

PCs, OCs: first dQ and the remaining principal components

Wₚ, Wₒ: the rotation matrix for PCs and OCs, respectively

mean_PCs: the mean of PCs before demeaned.

PCs are demeaned.

TermStructureModels.calibrate_μϕ_const — Method

calibrate_μϕ_const(μkQ_infty, σkQ_infty, ν0, yields, macros, τₙ, p; μϕ_const_PCs=[], medium_τ=collect(30:2:48), iteration=1000, data_scale=1200, medium_τ_pr=[], τ=[])

The purpose of the function is to calibrate a prior mean of the first dQ constant terms in our VAR. Adjust your prior setting based on the prior samples in outputs.

Input

μϕ_const_PCs is your prior mean of the first dQ constants. Our default option set it as a zero vector.

_ iteration is the number of prior samples.

τ::scalar is a maturity for calculating the constant part in the term premium.

If τ is empty, the function does not sampling the prior distribution of the constant part in the term premium.

Output(2)

priorλₚ, priorTP

samples from the prior distribution of λₚ

prior samples of constant part in the τ-month term premium

TermStructureModels.conditional_forecasts — Method

conditional_forecasts(S::Vector, τ, horizon, saved_θ, yields, macros, τₙ; mean_macros::Vector=[], data_scale=1200)

Input

scenarios, a result of the posterior sampler, and data

S[t] = conditioned scenario at time size(yields, 1)+t.

If we need an unconditional prediction, S = [].

If you are conditionaing a scenario, I assume S = Vector{Scenario}.

τ is a vector of maturities that term premiums of interest has.

horizon: maximum length of the predicted path. It should not be small than length(S).

saved_θ: the first output of function posterior_sampler.

mean_macros::Vector: If you demeaned macro variables, you can input the mean of the macro variables. Then, the output will be generated in terms of the un-demeaned macro variables.

Output

Vector{Forecast}(, iteration)

t'th rows in predicted yields, predicted factors, and predicted TP are the corresponding predicted value at time size(yields, 1)+t.

Mathematically, it is a posterior samples from future observation|past observation,scenario.

TermStructureModels.dcurvature_dτ — Method

dcurvature_dτ(τ; κQ)

This function calculate the first derivative of the curvature factor loading w.r.t. the maturity.

Input

κQ: The decay parameter

τ: The maturity that the derivative is calculated

Output

the first derivative of the curvature factor loading w.r.t. the maturity

TermStructureModels.dimQ — Method

dimQ()

It returns the dimension of Q-dynamics.

TermStructureModels.erase_nonstationary_param — Method

erase_nonstationary_param(saved_θ)

It filters out posterior samples that implies an unit root VAR system. Only stationary posterior samples remain.

Input

saved_θ is the first output of function posterior_sampler.

Output(2):

stationary samples, acceptance rate(%)

The second output indicates how many posterior samples remain.

TermStructureModels.fitted_YieldCurve — Method

fitted_YieldCurve(τ0, saved_Xθ::Vector{LatentSpace}; data_scale=1200)

It generates a fitted yield curve.

Input

τ0 is a set of maturities of interest. τ0 does not need to be the same as the one used for the estimation.

saved_Xθ is a transformed posterior sample using function latentspace.

Output

Vector{YieldCurve}(,# of iteration)

yields and latents contain initial observations.

TermStructureModels.generative — Method

generative(T, dP, τₙ, p, noise::Float64; κQ, kQ_infty, KₚXF, GₚXFXF, ΩXFXF, data_scale=1200)

This function generate a simulation data given parameters. Note that all parameters are the things in the latent factor state space (that is, parameters in struct LatentSpace). There is some differences in notations because it is hard to express mathcal letters in VScode. So, mathcal{F} in my paper is expressed in F in the VScode. And, "F" in my paper is expressed as XF.

Input:

noise = variance of the measurement errors

Output(3)

yields, latents, macros

yields = Matrix{Float64}(obs,T,length(τₙ))

latents = Matrix{Float64}(obs,T,dimQ())

macros = Matrix{Float64}(obs,T,dP - dimQ())

TermStructureModels.isstationary — Method

isstationary(GₚFF)

It checks whether a reduced VAR matrix has unit roots. If there is at least one unit root, return is false.

Input

GₚFF should not include intercepts. Also, GₚFF is dP by dP*p matrix that the coefficient at lag 1 comes first, and the lag p slope matrix comes last.

Output

boolean

TermStructureModels.latentspace — Method

latentspace(saved_θ, yields, τₙ; data_scale=1200)

This function translates the principal components state space into the latent factor state space.

Input

data_scale::scalar: In typical affine term structure model, theoretical yields are in decimal and not annualized. But, for convenience(public data usually contains annualized percentage yields) and numerical stability, we sometimes want to scale up yields, so want to use (data_scale*theoretical yields) as variable yields. In this case, you can use data_scale option. For example, we can set data_scale = 1200 and use annualized percentage monthly yields as yields.

Output

Vector{LatentSpace}(, iteration)

latent factors contain initial observations.

TermStructureModels.log_marginal — Method

log_marginal(PCs, macros, ρ, tuned::Hyperparameter, τₙ, Wₚ; ψ=[], ψ0=[], medium_τ, medium_τ_pr, fix_const_PC1)

This file calculates a value of our marginal likelihood. Only the transition equation is used to calculate it.

Input

tuned is a point where the marginal likelihood is evaluated.

ψ0 and ψ are multiplied with prior variances of coefficients of the intercept and lagged regressors in the orthogonalized transition equation. They are used for imposing zero prior variances. A empty default value means that you do not use this function. [ψ0 ψ][i,j] is corresponds to ϕ[i,j].

Output

the log marginal likelihood of the VAR system.

TermStructureModels.loglik_mea — Method

loglik_mea(yields, τₙ; κQ, kQ_infty, ϕ, σ²FF, Σₒ, data_scale)

This function generate a log likelihood of the measurement equation.

Output

the measurement equation part of the log likelihood

TermStructureModels.loglik_tran — Method

loglik_tran(PCs, macros; ϕ, σ²FF)

It calculate log likelihood of the transition equation.

Output

log likelihood of the transition equation.

TermStructureModels.posterior_sampler — Method

posterior_sampler(yields, macros, τₙ, ρ, iteration, tuned::Hyperparameter; medium_τ=collect(30:2:48), init_param=[], ψ=[], ψ0=[], γ_bar=[], medium_τ_pr=[], μkQ_infty=0, σkQ_infty=0.1, fix_const_PC1=false, data_scale=1200)

This is a posterior distribution sampler.

Input

iteration: # of posterior samples

tuned: optimized hyperparameters used during estimation

init_param: starting point of the sampler. It should be a type of Parameter.

Output(2)

Vector{Parameter}(posterior, iteration), acceptance rate of the MH algorithm

TermStructureModels.prior_κQ — Method

prior_κQ(medium_τ, pr)

The function derive the maximizer decay parameter κQ that maximize the curvature factor loading at each candidate medium-term maturity. And then, it impose a discrete prior distribution on the maximizers with a prior probability vector pr.

Input

medium_τ::Vector(candidate medium maturities, # of candidates)

pr::Vector(probability, # of candidates)

Output

discrete prior distribution that has a support of the maximizers κQ

TermStructureModels.reducedform — Method

reducedform(saved_θ, yields, macros, τₙ; data_scale=1200)

It converts posterior samples in terms of the reduced form VAR parameters.

Input

saved_θ is the first output of function posterior_sampler.

Output

Posterior samples in terms of struct ReducedForm

TermStructureModels.scenario_analysis — Method

scenario_analysis(S::Vector, τ, horizon, saved_θ, yields, macros, τₙ; mean_macros::Vector=[], data_scale=1200)

Input

scenarios, a result of the posterior sampler, and data

S[t] = conditioned scenario at time size(yields, 1)+t.

Set S = [] if you need an unconditional prediction.

If you are conditionaing a scenario, I assume S = Vector{Scenario}.

τ is a vector of maturities that term premiums of interest has.

horizon: maximum length of the predicted path. It should not be small than length(S).

saved_θ: the first output of function posterior_sampler.

mean_macros::Vector: If you demeaned macro variables, you can input the mean of the macro variables. Then, the output will be generated in terms of the un-demeaned macro variables.

Output

Vector{Forecast}(, iteration)

t'th rows in predicted yields, predicted factors, and predicted TP are the corresponding predicted value at time size(yields, 1)+t.

Mathematically, it is a posterior distribution of E[future obs|past obs, scenario, parameters].

TermStructureModels.term_premium — Method

term_premium(τ, τₙ, saved_θ, yields, macros; data_scale=1200)

This function generates posterior samples of the term premiums.

Input

maturity of interest τ for Calculating TP

saved_θ from function posterior_sampler

Output

Vector{TermPremium}(, iteration)

Outputs exclude initial observations.

TermStructureModels.tuning_hyperparameter — Method

tuning_hyperparameter(yields, macros, τₙ, ρ; populationsize=50, maxiter=10_000, medium_τ=collect(30:2:48), upper_q=[1 1; 1 1; 10 10; 100 100], μkQ_infty=0, σkQ_infty=0.1, upper_ν0=[], μϕ_const=[], fix_const_PC1=false, upper_p=18, μϕ_const_PC1=[], data_scale=1200, medium_τ_pr=[], init_ν0=[])

It optimizes our hyperparameters by maximizing the marginal likelhood of the transition equation. Our optimizer is a differential evolutionary algorithm that utilizes bimodal movements in the eigen-space(Wang, Li, Huang, and Li, 2014) and the trivial geography(Spector and Klein, 2006).

Input

When we compare marginal likelihoods between models, the data for the dependent variable should be the same across the models. To achieve that, we set a period of dependent variable based on upperp. For example, if upperp = 3, yields[4:end,:] and macros[4:end,:] are the data for our dependent variable. yields[1:3,:] and macros[1:3,:] are used for setting initial observations for all lags.

populationsize and maxiter is a option for the optimizer.

The lower bounds for q and ν0 are 0 and dP+2.

The upper bounds for q, ν0 and VAR lag can be set by upper_q, upper_ν0, upper_p.

Our default option for upper_ν0 is the time-series length of the data.

If you use our default option for μϕ_const,

μϕ_const[dQ+1:end] is a zero vector.

μϕ_const[1:dQ] is calibrated to make a prior mean of λₚ a zero vector.

After step 2, μϕ_const[1] is replaced with μϕ_const_PC1 if it is not empty.

μϕ_const = Matrix(your prior, dP, upper_p)

μϕ_const[:,i] is a prior mean for the VAR(i) constant. Therefore μϕconst is a matrix only in this function. In other functions, `μϕconst` is a vector for the orthogonalized VAR system with your selected lag.

When fix_const_PC1==true, the first element in a constant term in our orthogonalized VAR is fixed to its prior mean during the posterior sampling.

data_scale::scalar: In typical affine term structure model, theoretical yields are in decimal and not annualized. But, for convenience(public data usually contains annualized percentage yields) and numerical stability, we sometimes want to scale up yields, so want to use (data_scale*theoretical yields) as variable yields. In this case, you can use data_scale option. For example, we can set data_scale = 1200 and use annualized percentage monthly yields as yields.

Output(2)

optimized Hyperparameter, optimization result

Be careful that we minimized the negative log marginal likelihood, so the second output is about the minimization problem.

TermStructureModels.ϕ_2_ϕ₀_C — Method

ϕ_2_ϕ₀_C(; ϕ)

It divide ϕ into the lagged regressor part and the contemporaneous regerssor part.

Output(3)

ϕ0, C = C0 + I, C0

ϕ0: coefficients for the lagged regressors

C: coefficients for the dependent variables when all contemporaneous variables are in the LHS of the orthogonalized equations. Therefore, the diagonals of C is ones. Note that since the contemporaneous variables get negative signs when they are at the RHS, the signs of C do not change whether they are at the RHS or LHS.

TermStructureModels.ϕ_σ²FF_2_ΩFF — Method

ϕ_σ²FF_2_ΩFF(; ϕ, σ²FF)

It construct ΩFF from statistical parameters.

Output

ΩFF

Internal Functions

TermStructureModels.Aₓ — Method

Aₓ(aτ_, τₙ)

Input

aτ_ is an output of function aτ.

Output

Aₓ

TermStructureModels.Aₚ — Method

Aₚ(Aₓ_, Bₓ_, T0P_, Wₒ)

Input

Aₓ_, Bₓ_, and T0P_ are outputs of function Aₓ, Bₓ, and T0P, respectively.

Output

Aₚ

TermStructureModels.Bₓ — Method

Bₓ(bτ_, τₙ)

Input

bτ_ is an output of function bτ.

Output

Bₓ

TermStructureModels.Bₚ — Method

Bₚ(Bₓ_, T1X_, Wₒ)

Input

Bₓ_ and T1X_ are outputs of function Bₓ and T1X, respectively.

Output

Bₚ

TermStructureModels.Kϕ — Method

Kϕ(i, V, Xϕ, dP)

TermStructureModels.Minnesota — Method

Minnesota(l, i, j; q, ν0, Ω0)

It return unscaled prior variance of the Minnesota prior.

Input

lag l, dependent variable i, regressor j in the VAR(p)

q[:,1] and q[:,2] are [own, cross, lag, intercept] shrikages for the first dQ and remaining dP-dQ equations, respectively.

ν0(d.f.), Ω0(scale): Inverse-Wishart prior for the error-covariance matrix of VAR(p).

Output

Minnesota part in the prior variance

TermStructureModels.NIG_NIG — Method

NIG_NIG(y, X, β₀, B₀, α₀, δ₀)

Normal-InverseGamma-Normal-InverseGamma update

prior: β|σ² ~ MvNormal(β₀,σ²B₀), σ² ~ InverseGamma(α₀,δ₀)

likelihood: y|β,σ² = Xβ + MvNormal(zeros(T,1),σ²I(T))

Output(2)

β, σ²

posterior sample

TermStructureModels.PCs_2_latents — Method

PCs_2_latents(yields, τₙ; κQ, kQ_infty, KₚF, GₚFF, ΩFF, data_scale)

Notation XF is for the latent factor space and notation F is for the PC state space.

Input

data_scale::scalar: In typical affine term structure model, theoretical yields are in decimal and not annualized. But, for convenience(public data usually contains annualized percentage yields) and numerical stability, we sometimes want to scale up yields, so want to use (data_scale*theoretical yields) as variable yields. In this case, you can use data_scale option. For example, we can set data_scale = 1200 and use annualized percentage monthly yields as yields.

Output(6)

latent, κQ, kQ_infty, KₚXF, GₚXFXF, ΩXFXF

latent factors contain initial observations.

TermStructureModels.S — Method

S(i; Ω0)

TermStructureModels.S_hat — Method

S_hat(i, m, V, yϕ, Xϕ, dP; Ω0)

TermStructureModels.T0P — Method

T0P(T1X_, Aₓ_, Wₚ, c)

Input

T1X_ and Aₓ_ are outputs of function T1X and Aₓ, respectively. c is a sample mean of PCs.

Output

T0P

TermStructureModels.T1X — Method

T1X(Bₓ_, Wₚ)

Input

Bₓ_ if an output of function Bₓ.

Output

T1X

TermStructureModels._conditional_forecasts — Method

_conditional_forecasts(S, τ, horizon, yields, macros, τₙ; κQ, kQ_infty, ϕ, σ²FF, Σₒ, mean_macros, data_scale)

TermStructureModels._scenario_analysis — Method

_scenario_analysis(S, τ, horizon, yields, macros, τₙ; κQ, kQ_infty, ϕ, σ²FF, Σₒ, mean_macros, data_scale)

TermStructureModels._scenario_analysis_unconditional — Method

_scenario_analysis_unconditional(τ, horizon, yields, macros, τₙ; κQ, kQ_infty, ϕ, σ²FF, Σₒ, mean_macros, data_scale)

TermStructureModels._termPremium — Method

_termPremium(τ, PCs, macros, bτ_, T0P_, T1X_; κQ, kQ_infty, KₚF, GₚFF, ΩPP, data_scale)

This function calculates a term premium for maturity τ.

Input

data_scale::scalar = In typical affine term structure model, theoretical yields are in decimal and not annualized. But, for convenience(public data usually contains annualized percentage yields) and numerical stability, we sometimes want to scale up yields, so want to use (data_scale*theoretical yields) as variable yields. In this case, you can use data_scale option. For example, we can set data_scale = 1200 and use annualized percentage monthly yields as yields.

Output(4)

TP, timevarying_TP, const_TP, jensen

TP: term premium of maturity τ

timevarying_TP: contributions of each [PCs macros] on TP at each time $t$ (row: time, col: variable)

const_TP: Constant part of TP

jensen: Jensen's Ineqaulity part in TP

Output excludes the time period for the initial observations.

TermStructureModels._unconditional_forecasts — Method

_unconditional_forecasts(τ, horizon, yields, macros, τₙ; κQ, kQ_infty, ϕ, σ²FF, Σₒ, mean_macros, data_scale)

TermStructureModels.aτ — Method

aτ(N, bτ_, τₙ, Wₚ; kQ_infty, ΩPP, data_scale)
aτ(N, bτ_; kQ_infty, ΩXX, data_scale)

The function has two methods(multiple dispatch).

Input

When Wₚ ∈ arguments: It calculates aτ using ΩPP.

Otherwise: It calculates aτ using ΩXX = ΩXFXF[1:dQ, 1:dQ], so parameters are in the latent factor space. So, we do not need Wₚ.

bτ_ is an output of function bτ.

data_scale::scalar: In typical affine term structure model, theoretical yields are in decimal and not annualized. But, for convenience(public data usually contains annualized percentage yields) and numerical stability, we sometimes want to scale up yields, so want to use (data_scale*theoretical yields) as variable yields. In this case, you can use data_scale option. For example, we can set data_scale = 1200 and use annualized percentage monthly yields as yields.

Output

Vector(Float64)(aτ,N)

For i'th maturity, Output[i] is the corresponding aτ.

TermStructureModels.bτ — Method

bτ(N; κQ)

It solves the difference equation for bτ.

Output

for maturity i, bτ[:, i] is a vector of factor loadings.

TermStructureModels.jensens_inequality — Method

jensens_inequality(τ, bτ_, T1X_; ΩPP, data_scale)

This function evaluate the Jensen's Ineqaulity term. All term is invariant with respect to the data_scale, except for this Jensen's inequality term. So, we need to scale down the term by data_scale.

Output

Jensen's Ineqaulity term for aτ of maturity τ.

TermStructureModels.post_kQ_infty — Method

post_kQ_infty(μkQ_infty, σkQ_infty, yields, τₙ; κQ, ϕ, σ²FF, Σₒ, data_scale)

Output

Full conditional posterior distribution

TermStructureModels.post_Σₒ — Method

post_Σₒ(yields, τₙ; κQ, kQ_infty, ΩPP, γ, p, data_scale)

Posterior sampler for the measurement errors

Output

Vector{Dist}(IG, N-dQ)

TermStructureModels.post_γ — Method

post_γ(; γ_bar, Σₒ)

Posterior sampler for the population measurement error

Output

Vector{Dist}(Gamma,length(Σₒ))

TermStructureModels.post_κQ — Method

post_κQ(yields, prior_κQ_, τₙ; kQ_infty, ϕ, σ²FF, Σₒ, data_scale)

Input

prior_κQ_ is a output of function prior_κQ.

Output

Full conditional posterior distribution

TermStructureModels.post_ϕ_σ²FF — Method

post_ϕ_σ²FF(yields, macros, μϕ_const, ρ, prior_κQ_, τₙ; ϕ, ψ, ψ0, σ²FF, q, ν0, Ω0, κQ, kQ_infty, Σₒ, fix_const_PC1, data_scale)

Full-conditional posterior sampler for ϕ and σ²FF

Input

prior_κQ_ is a output of function prior_κQ.

When fix_const_PC1==true, the first element in a constant term in our orthogonalized VAR is fixed to its prior mean during the posterior sampling.

Output(3)

ϕ, σ²FF, isaccept=Vector{Bool}(undef, dQ)

It gives a posterior sample.

TermStructureModels.prior_C — Method

prior_C(; Ω0::Vector)

We translate the Inverse-Wishart prior to a series of the Normal-Inverse-Gamma (NIG) prior distributions. If the dimension is dₚ, there are dₚ NIG prior distributions. This function generates Normal priors.

Output:

unscaled prior of C in the LDLt decomposition, ΩFF = inv(C)*diagm(σ²FF)*inv(C)'

Important note

prior variance for C[i,:] = σ²FF[i]*variance of output[i,:]

TermStructureModels.prior_γ — Method

prior_γ(yields, p)

There is a hierarchcal structure in the measurement equation. The prior means of the measurement errors are γ[i] and each γ[i] follows Gamma(1,γ_bar) distribution. This function decides γ_bar empirically. OLS is used to estimate the measurement equation and then a variance of residuals is calculated for each maturities. An inverse of the average residual variances is set to γ_bar.

Output

hyperparameter γ_bar

TermStructureModels.prior_σ²FF — Method

prior_σ²FF(; ν0, Ω0::Vector)

We translate the Inverse-Wishart prior to a series of the Normal-Inverse-Gamma (NIG) prior distributions. If the dimension is dₚ, there are dₚ NIG prior distributions. This function generates Inverse-Gamma priors.

Output:

prior of σ²FF in the LDLt decomposition,ΩFF = inv(C)*diagm(σ²FF)*inv(C)'

Each element in the output follows Inverse-Gamma priors.

TermStructureModels.prior_ϕ0 — Method

prior_ϕ0(μϕ_const, ρ::Vector, prior_κQ_, τₙ, Wₚ; ψ0, ψ, q, ν0, Ω0, fix_const_PC1)

This part derives the prior distribution for coefficients of the lagged regressors in the orthogonalized VAR.

Input

prior_κQ_ is a output of function prior_κQ.

When fix_const_PC1==true, the first element in a constant term in our orthogonalized VAR is fixed to its prior mean during the posterior sampling.

Output

Normal prior distributions on the slope coefficient of lagged variables and intercepts in the orthogonalized equation.

Output[:,1] for intercepts, Output[:,1+1:1+dP] for the first lag, Output[:,1+dP+1:1+2*dP] for the second lag, and so on.

Important note

prior variance for ϕ[i,:] = σ²FF[i]*var(output[i,:])

TermStructureModels.yϕ_Xϕ — Method

yϕ_Xϕ(PCs, macros, p)

This function generate the dependent variable and the corresponding regressors in the orthogonalized transition equation.

Output(4)

yϕ, Xϕ = [ones(T - p) Xϕ_lag Xϕ_contemporaneous], [ones(T - p) Xϕ_lag], Xϕ_contemporaneous

yϕ and Xϕ is a full matrix. For i'th equation, the dependent variable is yϕ[:,i] and the regressor is Xϕ.

Xϕ is same to all orthogonalized transtion equations. The orthogonalized equations are different in terms of contemporaneous regressors. Therefore, the corresponding regressors in Xϕ should be excluded. The form of parameter ϕ do that task by setting the coefficients of the excluded regressors to zeros. In particular, for last dP by dP block in ϕ, the diagonals and the upper diagonal elements should be zero.

TermStructureModels.ν — Method

ν(i, dP; ν0)

TermStructureModels.ϕ_hat — Method

ϕ_hat(i, m, V, yϕ, Xϕ, dP)

TermStructureModels.ϕ_σ²FF_2_ΩPP — Method

ϕ_σ²FF_2_ΩPP(; ϕ, σ²FF)

It construct ΩPP from statistical parameters.

Output

ΩPP